Move IKeyValueStore.h to fdbserver/kvstore
