+++ fdbserver/CMakeLists.txt
@@ -25,8 +25,8 @@ if(WITH_ROCKSDB)
 endif()
 
 add_subdirectory(core)
-add_subdirectory(logsystem)
 add_subdirectory(kvstore)
+add_subdirectory(logsystem)
 add_subdirectory(mocks3)
 add_subdirectory(clustercontroller)
 add_subdirectory(backupworker)
