+++ fdbserver/clustercontroller/CMakeLists.txt
@@ -17,4 +17,4 @@ target_include_directories(fdbserver_clustercontroller
     ${CMAKE_CURRENT_BINARY_DIR}
     ${CMAKE_SOURCE_DIR}/fdbserver/datadistributor/include
     ${CMAKE_BINARY_DIR}/fdbserver/datadistributor/include)
-target_link_libraries(fdbserver_clustercontroller PRIVATE fdbserver_core fdbserver_logsystem)
+target_link_libraries(fdbserver_clustercontroller PRIVATE fdbserver_core fdbserver_kvstore fdbserver_logsystem)
