+++ fdbserver/commitproxy/CommitProxyServer.actor.cpp
@@ -42,7 +42,7 @@
 #include "fdbserver/logsystem/ApplyMetadataMutation.h"
 #include "fdbserver/core/ConflictBatch.h"
 #include "fdbserver/core/DataDistributorInterface.h"
-#include "fdbserver/core/IKeyValueStore.h"
+#include "fdbserver/kvstore/IKeyValueStore.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/logsystem/LogSystem.h"
 #include "fdbserver/kvstore/FDBExecHelper.h"
