+++ fdbserver/datadistributor/include/fdbserver/datadistributor/DDTeamCollection.h
@@ -32,7 +32,6 @@
 #include "fdbclient/ManagementAPI.h"
 #include "fdbclient/RunRYWTransaction.h"
 #include "fdbrpc/Replication.h"
-#include "fdbserver/core/IKeyValueStore.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/core/MoveKeys.h"
 #include "fdbserver/core/TLogInterface.h"
