+++ fdbserver/kvstore/KeyValueStoreRocksDB.actor.cpp
@@ -66,7 +66,7 @@
 #endif // WITH_ROCKSDB
 
 #include "fdbserver/core/Knobs.h"
-#include "fdbserver/core/IKeyValueStore.h"
+#include "fdbserver/kvstore/IKeyValueStore.h"
 #include "fdbserver/core/RocksDBCheckpointUtils.h"
 #include "RocksDBCommon.h"
