+++ fdbserver/kvstore/KeyValueStoreShardedRocksDB.actor.cpp
@@ -45,7 +45,7 @@
 #endif // WITH_ROCKSDB
 
 #include "fdbserver/core/Knobs.h"
-#include "fdbserver/core/IKeyValueStore.h"
+#include "fdbserver/kvstore/IKeyValueStore.h"
 #include "fdbserver/core/RocksDBCheckpointUtils.h"
 #include "RocksDBCommon.h"
 #include "flow/actorcompiler.h" // has to be last include
