+++ fdbserver/storageserver/storageserver.actor.cpp
@@ -42,7 +42,7 @@
 #include "fdbserver/core/BulkDumpUtil.h"
 #include "fdbserver/core/BulkLoadUtil.h"
 #include "fdbserver/core/FDBRocksDBVersion.h"
-#include "fdbserver/core/IKeyValueStore.h"
+#include "fdbserver/kvstore/IKeyValueStore.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/kvstore/FDBExecHelper.h"
 #include "fdbserver/core/LatencyBandConfig.h"
