+++ fdbserver/tlog/TLogServer.actor.cpp
@@ -26,7 +26,7 @@
 #include "fdbclient/RunRYWTransaction.h"
 #include "fdbclient/SystemData.h"
 #include "fdbclient/FDBTypes.h"
-#include "fdbserver/core/IKeyValueStore.h"
+#include "fdbserver/kvstore/IKeyValueStore.h"
 #include "fdbserver/core/IDiskQueue.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/core/MutationTracking.h"
