+++ fdbserver/tlog/TestTLogServer.h
@@ -26,7 +26,7 @@
 
 #include "fdbclient/FDBTypes.h"
 #include "fdbserver/core/IDiskQueue.h"
-#include "fdbserver/core/IKeyValueStore.h"
+#include "fdbserver/kvstore/IKeyValueStore.h"
 #include "fdbserver/logsystem/LogSystem.h"
 #include "fdbserver/core/ResolverInterface.h"
 #include "fdbserver/core/TLogInterface.h"
