+++ fdbserver/workloads/StorageServerCheckpointRestoreTest.cpp
@@ -21,7 +21,7 @@
 #include "fdbclient/ManagementAPI.h"
 #include "fdbclient/NativeAPI.actor.h"
 #include "fdbrpc/simulator.h"
-#include "fdbserver/core/IKeyValueStore.h"
+#include "fdbserver/kvstore/IKeyValueStore.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/core/ServerCheckpoint.h"
 #include "fdbserver/core/MoveKeys.h"
